Toggle side menu on actions
