+++ src/components/MainApp.tsx
@@ -1,10 +1,8 @@
 import { Layout } from "antd";
 import React from "react";
-import { useDispatch } from "react-redux";
 
 import { useMediaQueryBreakpoints } from "../hooks/useMediaQueryBreakpoints";
 import { useAppSelector } from "../store/store";
-import { ActionTypes } from "../store/types";
 import colors from "../styles/colors";
 import ContentWrapper from "./ContentWrapper";
 import Footer from "./Footer/Footer";
@@ -13,15 +11,9 @@ import OverlaySearchResults from "./Search/OverlaySearchResults";
 import Sidebar from "./Sidebar";
 
 const MainApp = ({ children }) => {
-    const dispatch = useDispatch();
-
-    const { enableOverlay, menuCollapsed } = useAppSelector((state) => {
+    const { enableOverlay } = useAppSelector((state) => {
         return {
             enableOverlay: state?.search?.overlayVisible,
-            menuCollapsed:
-                state?.menuCollapsed !== undefined
-                    ? state?.menuCollapsed
-                    : true,
         };
     });
 
@@ -30,15 +22,7 @@ const MainApp = ({ children }) => {
 
     return (
         <Layout style={{ minHeight: "100vh" }}>
-            <Sidebar
-                menuCollapsed={menuCollapsed}
-                onToggleSidebar={() => {
-                    dispatch({
-                        type: ActionTypes.TOGGLE_MENU,
-                        menuCollapsed: !menuCollapsed,
-                    });
-                }}
-            />
+            <Sidebar />
             <Layout style={{ background: colors.white }}>
                 <Header />
                 <ContentWrapper>{children}</ContentWrapper>
