+++ src/components/Sidebar.tsx
@@ -1,15 +1,29 @@
 import { Drawer } from "antd";
 import React from "react";
+import { useDispatch } from "react-redux";
+import actions from "../store/actions";
+import { useAppSelector } from "../store/store";
 
 import colors from "../styles/colors";
 import AletheiaMenu from "./AletheiaMenu";
 import Logo from "./Header/Logo";
 
-const Sidebar = ({ menuCollapsed, onToggleSidebar }) => {
+const Sidebar = () => {
+    const dispatch = useDispatch();
+
+    const { menuCollapsed } = useAppSelector((state) => {
+        return {
+            menuCollapsed:
+                state?.menuCollapsed !== undefined
+                    ? state?.menuCollapsed
+                    : true,
+        };
+    });
+
     return (
         <Drawer
             visible={!menuCollapsed}
-            onClose={onToggleSidebar}
+            onClose={() => dispatch(actions.closeSideMenu())}
             width="17rem"
             placement="left"
             bodyStyle={{ padding: 0 }}
